Prevent package update report creation when none.
Do not create a report when there are no package updates.

+++ pkg-upgrade.sh
@@ -5,8 +5,8 @@ THIS_FILE="pkg-upgrade.sh"
 #
 # Brief Description
 #
-# Usage: bash pkg-upgrade.sh
-#        (not sh pkg-upgrade.sh)
+# Usage: bash osup.sh
+#        (not sh osup.sh)
 #
 # Script osup.sh will show a description of each upgradeable package before
 # upgrading each package.
@@ -15,7 +15,9 @@ THIS_FILE="pkg-upgrade.sh"
 ##
 ## Code Change History
 ##
-## 2019-03-28 *Adjusted displayed package descriptions format.
+## 2019-03-28 *Adjusted displayed package descriptions format. 
+##            *Added message if there are no packages to update.
+##            *if there are no packages to update, do not create a report.
 ##
 ## 2019-03-26 *Added detection of installed file viewer.
 ##
@@ -26,29 +28,35 @@ sudo apt list --upgradable > uplist.tmp  # Raw data output from command, "apt li
 awk -F / '{ print $1 }' uplist.tmp > uplist2.tmp  # Parse raw data to show each package title only.
 sed -i s/Listing...// uplist2.tmp  # Delete string "Listing...Done" from list of packages.
 awk 'NF' uplist2.tmp > uplist.tmp  # Delete empty line left-over from deleting string.
-TITLE="***Description of upgradeable packages***"
-echo $TITLE > uplist2.tmp
-#
-echo $TITLE
-echo
-while read XSTR
-do
-      echo >> uplist2.tmp
-      echo $XSTR >> uplist2.tmp
-      apt-cache show $XSTR | grep Description --max-count=1 --after-context=2 >> uplist2.tmp
-done < uplist.tmp
-# Detect installed file viewer.
-RUNAPP=0
-for FILEVR in most more less
-do
-    if [ $RUNAPP -eq 0 ] ; then
-       type $FILEVR >/dev/null 2>&1  # Test if $FILEVR application is installed.
-       ERROR=$?
-       if [ $ERROR -eq 0 ] ; then
-          $FILEVR uplist2.tmp
-          RUNAPP=1
+if [ -s uplist.tmp ] ; then  # If tmp file has data (list of packages to be updated).
+   TITLE="***Description of upgradeable packages***"
+   echo $TITLE > uplist2.tmp
+   #
+   echo $TITLE
+   echo
+   # Extract the "Description" from the raw data output from command "apt-cache show <pkg name>". 
+   while read XSTR
+   do
+         echo >> uplist2.tmp
+         echo $XSTR >> uplist2.tmp
+         apt-cache show $XSTR | grep Description --max-count=1 --after-context=2 >> uplist2.tmp
+   done < uplist.tmp
+   # Detect installed file viewer.
+   RUNAPP=0
+   for FILEVR in most more less
+       do
+       if [ $RUNAPP -eq 0 ] ; then
+          type $FILEVR >/dev/null 2>&1  # Test if $FILEVR application is installed.
+          ERROR=$?
+          if [ $ERROR -eq 0 ] ; then
+             $FILEVR uplist2.tmp
+             RUNAPP=1
+          fi
        fi
-    fi
-done
-sudo apt upgrade
-rm uplist*.tmp
+   done
+   sudo apt upgrade
+else
+   echo
+   echo "No packages to update. All packages are up to date."
+   #rm uplist*.tmp
+fi
